fix 'size<vector...>' for no-CTPS compilers
git-svn-id: http://svn.boost.org/svn/boost/trunk@18786 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/mpl/vector/aux_/numbered.hpp
@@ -167,7 +167,7 @@ template<>
 struct size_traits< aux::vector_tag<i> >
 {
     template< typename Vector > struct algorithm
-        : integral_c<int,i>
+        : integral_c<long,i>
     {
     };
 };
